Create FITS2xyz
FITS format of image of Kuiper belt objects or so and is read in z,x,y format where z is the number of bands!! It can be pretty confusing for users who are used to of x,y,z format! This code rearranges the format to x,y,z for ease!!! Also converts all the negative values to 0.